+++ test/component/composer/cases/case89.js
@@ -1,6 +1,5 @@
-export const title       = '55) Error: g-content expression';
-export const description = 'launch an error expression in g-content ';
-
+export const title       = '89) Multiple errors';
+export const description = 'launch plugin and render errors';
 
 export function script () {
   const result    = document.querySelector('#result');
@@ -13,14 +12,20 @@ export function script () {
   } else {
     show();
   }
+  component.addEventListener('error', show);
+  component.data.value="hello";
+  setTimeout(() =>component.data.value="hello", 200);
 }
 
 export default `
 <div id="container">
-  <g-composer>
+  <g-composer value="0">
     <svg viewBox="0 0 100 100">
-      <text x="0" y="0" g-content="wrong"></text>
+      <rect :x="value * 2"
+            @click="wrong()" 
+            y="0" width="100" height="100" fill="red"/>
     </svg>
+    <script type="plugin" src="./non-exist.js"></script>
   </g-composer>
 </div>
 <pre id="result"></pre>